+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - use heroicons/react icons
 - components: SearchInput: auto "submit" form when user types (debounced) vs on actual submit (hitting enter key)
 - components: UserLinkCard, udpated styles, added location
+- pages: home: refactored to use SearchUser widget
 #### Fixed
 - search: filter organizations (empty objects) from search results